Move EM27 sensors and DECADES code into new `sensors` module
