+++ finesse/gui/hardware_set/dummy.yaml
@@ -9,6 +9,6 @@ devices:
   temperature_monitor:
     class_name: temperature.dummy_temperature_monitor.DummyTemperatureMonitor
   em27_sensors:
-    class_name: spectrometer.dummy_em27_sensors.DummyEM27Sensors
+    class_name: sensors.dummy_em27_sensors.DummyEM27Sensors
   spectrometer:
     class_name: spectrometer.dummy_opus_interface.DummyOPUSInterface
