+++ finesse/gui/hardware_set/finesse_seneca.yaml
@@ -21,7 +21,7 @@ devices:
       port: "0403:6001"
       baudrate: 57600
   em27_sensors:
-    class_name: spectrometer.em27_sensors.EM27Sensors
+    class_name: sensors.em27_sensors.EM27Sensors
   spectrometer:
     class_name: spectrometer.opus_interface.OPUSInterface
     params:
