--- finesse/hardware/plugins/decades/__init__.py
@@ -1,4 +1,0 @@
-"""Code for interfacing with the DECADES server.
-
-This module contains code for polling aircraft parameter values using the DECADES API.
-"""
