+++ finesse/hardware/plugins/sensors/__init__.py
@@ -1,0 +1,1 @@
+"""This module contains code for devices yielding sensor readings."""
