+++ finesse/hardware/plugins/sensors/decades.py
