+++ finesse/hardware/plugins/sensors/diag_autom.htm
