+++ finesse/hardware/plugins/sensors/em27_sensors.py
